+++ src/RageTimer.cpp
@@ -122,7 +122,7 @@ RageTimer RageTimer::Sum(const RageTimer &lhs, float tm)
 {
 	/* tm == 5.25  -> secs =  5, us = 5.25  - ( 5) = .25
 	 * tm == -1.25 -> secs = -2, us = -1.25 - (-2) = .75 */
-	int seconds = (int) floorf(tm);
+	int seconds = static_cast<int>(std::floor(tm));
 	int us = int( (tm - seconds) * TIMESTAMP_RESOLUTION );
 
 	RageTimer ret(0,0); // Prevent unnecessarily checking the time
